ryno-v2-scripts/requirements.txt
fastapi==0.95.0
numpy==1.23.5
openai>=1.10.0,<2.0.0
pdfplumber==0.8.0
pinecone-client>=3,<4  # Updated line
protobuf==4.22.1
pydantic>=1.9,<2.0.0
python-dotenv==1.0.0
requests==2.28.2
uvicorn==0.21.1
werkzeug
google-cloud-storage
langchain-pinecone 
langchain-openai 
langchain-core
langchain
